PlaceViewer 쓸데없는 기능 없애면서 RoomViewer로 변경.
