+++ ProjectSettings/TagManager.asset
@@ -13,7 +13,7 @@ TagManager:
   - UI
   - 
   - 
-  - Place Viewed
+  - Room Viewed
   - 
   - 
   -
